+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Features
 - Add `StudioClient` as connector to PhariaStudio for submitting traces.
 - You can now specify a `chunk_overlap` when creating an index in the Document Index.
+- Add support for monitoring progress in the document index connector when embedding documents.
 
 ### Fixes
 ...